pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[project]
name = "scheme-assistant"
version = "0.1.0"
description = "Multi-agent scheme recommendation system using Google ADK"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.1.0",
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-discoveryengine>=0.11.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]